Create README.md
Welcome to the Pokedex-Challenge wiki!

This Pokedex was created to deepening in React, to show and learn my knowledge and new JS libraries. By clicking on this code, you are going to see a lot of my own creation and some support in searches from other React developers. If you need some help or any support, I'll be here for you.

((*This React App was created using PokeAPI. *))

Enjoy!